Code cleanup.

+++ jhotdraw8/src/main/java/org.jhotdraw8.application/org/jhotdraw8/collection/ReadOnlyListIterator.java
@@ -100,7 +100,7 @@ public class ReadOnlyListIterator<E> implements Iterator<E>, ListIterator<E>, Sp
 
     @Override
     public long estimateSize() {
-        return (long) (size - index);
+        return size - index;
     }
 
     @Override
